+++ ochre/utils/schedule.py
@@ -65,6 +65,9 @@ SCHEDULE_NAMES = {
         "Power Outage": None,
         "No Space Heating": None,
         "No Space Cooling": None,
+        "electric_vehicle": None,
+        "electric_vehicle_charging": None,
+        "electric_vehicle_discharging": None,
     },
 }